Former-commit-id: 8543842f16e5ae70a57ebda171239df0b78ee7b8 [formerly e7ec80c0ca1887c22ef456a57c8aecc3dd8d583e] [formerly a97734c786899b7531eebad8320e1dce513244f2] [formerly a97734c786899b7531eebad8320e1dce513244f2 [formerly 21e3eeca05fd0f2078fa6f6ccf5ce8ea60d543fd]] [formerly 8543842f16e5ae70a57ebda171239df0b78ee7b8 [formerly e7ec80c0ca1887c22ef456a57c8aecc3dd8d583e] [formerly a97734c786899b7531eebad8320e1dce513244f2] [formerly a97734c786899b7531eebad8320e1dce513244f2 [formerly 21e3eeca05fd0f2078fa6f6ccf5ce8ea60d543fd]] [formerly 86c0fe78f27f766eb282b1a49781c6f34ef4798f [formerly a97734c786899b7531eebad8320e1dce513244f2 [formerly 21e3eeca05fd0f2078fa6f6ccf5ce8ea60d543fd] [formerly 86c0fe78f27f766eb282b1a49781c6f34ef4798f [formerly 6ec350a33b6f5bca31906ff2f7b5b76911f33973]]]]]
Former-commit-id: 86c0fe78f27f766eb282b1a49781c6f34ef4798f
Former-commit-id: 9e709f297ce239dca33ae624d2c1d9ac03db3e2e [formerly 6264fd681d4408ec0347652e6b2a03bbfdb0cbbd] [formerly 374f012e85c976568392caecbe5d796e9d8b198a] [formerly 6b7d3d1576638cc5e2c537430910f2fae97be9ca [formerly 277e90770a5c5453a1ebffc7aa107e33b2fb38a6] [formerly 374f012e85c976568392caecbe5d796e9d8b198a [formerly c318aa65479ca5c3f6bf5291c99afcd1ee7e747b]]]
Former-commit-id: 8ad7af7ca00518cd5dc66350c3fb2b2669b2a858 [formerly ad44818ed2b68b4354be44e373ce58805e068972] [formerly 93d29b05a98ea8648e6194c493d2fec270870f86 [formerly 8eb249f8669ec07bccb571eb04614dd72c3665a7]]
Former-commit-id: 93d29b05a98ea8648e6194c493d2fec270870f86
Former-commit-id: 350eb0a1ca1740db4eea151504ec33d45cb14890

+++ edexOsgi/build.edex/opt/db/ddl/maps/importShapeFile.sh
@@ -17,10 +17,19 @@
 # 
 # See the AWIPS II Master Rights File ("Master Rights File.pdf") for
 # further licensing information.
+#
+#
+# SOFTWARE HISTORY
+# Date         Ticket#    Engineer    Description
+# ------------ ---------- ----------- --------------------------
+# 03/25/2014    #2664     randerso    Added support for importing non-WGS84 shape files
+#     
 ##
-if [ $# -lt 3 ] ; then
+
+function usage()
+{
     echo
-    echo usage: `basename $0` shapefile schema table [simplev [dbUser [dbPort [installDir]]]]
+    echo usage: `basename $0` shapefile schema table [simplev [dbUser [dbPort [installDir [srid]]]]]
     echo "where: shapefile - pathname of the shape file to be imported"
     echo "       schema    - database schema where the shape file is to be imported"
     echo "       table     - database table where the shape file is to be imported"
@@ -28,11 +37,22 @@ if [ $# -lt 3 ] ; then
     echo "       dbUser    - optional database user id"
     echo "       dbPort    - optional database port number"
     echo "       installDir- optional directory path to awips installation"
-    echo "example: `basename $0` uscounties.shp mapdata County 0.064,0.016,0.004,0.001 awips 5432 /awips2"
+    echo "       srid      - optional srid of source shape file"
+    echo "example: `basename $0` uscounties.shp mapdata County 0.064,0.016,0.004,0.001 awips 5432 /awips2" 
+}
+
+if [ $# -lt 3 ] ; then
+    usage
     exit -1
 fi
 
-SHAPEFILEPATH=${1}
+SHP_PATH=${1}
+SHP_DIR="${SHP_PATH%/*}"    # shape file dir
+SHP_NAME="${SHP_PATH##*/}"  # shape file name with extension
+SHP_BASE="${SHP_NAME%.*}"   # shape file name without extension
+SHP_EXT="${SHP_NAME##*.}"   # shape file extension
+PRJ_PATH="${SHP_DIR}/${SHP_BASE}.prj"
+
 SCHEMA=`echo "${2}" | tr '[:upper:]' '[:lower:]'`
 TABLE=`echo "${3}" | tr '[:upper:]' '[:lower:]'`
 SIMPLEVS=${4}
@@ -57,16 +77,38 @@ else
     PSQLBINDIR=${7}/psql/bin/
 fi
 
-SHAPEFILENAME=`basename ${SHAPEFILEPATH}`
-echo "  Importing ${SHAPEFILENAME} into ${SCHEMA}.${TABLE} ..."
+if [ -z $8 ] ; then
+    if [ -e $PRJ_PATH ]
+    then
+        echo
+        echo "WARNING, found projection file: ${PRJ_PATH}"
+        echo "It is probable that this shape file is not in EPSG:4326 (WGS 84, unprojected lat/lon) format."
+        echo "Please determine the correct srid by uploading the .prj file to http://prj2epsg.org/search"
+        echo "and re-run `basename $0` supplying the correct srid."
+        usage
+        exit -1
+    fi
+
+    SRID=4326
+else
+    SRID=${8}:4326
+fi
+
+if [ ! -r ${SHP_PATH} ]; then
+    echo
+    echo "ERROR, file not found or is not readable: ${SHP_PATH}"
+    exit -1
+fi
+
+echo "  Importing ${SHP_NAME} into ${SCHEMA}.${TABLE} ..."
 ${PSQLBINDIR}psql -d maps -U ${PGUSER} -q -p ${PGPORT} -c "
     DELETE FROM public.geometry_columns WHERE f_table_schema = '${SCHEMA}' AND f_table_name = '${TABLE}';
-    DELETE from ${SCHEMA}.map_version WHERE table_name='${TABLE}';
+    DELETE FROM ${SCHEMA}.map_version WHERE table_name='${TABLE}';
     DROP TABLE IF EXISTS ${SCHEMA}.${TABLE}
 "
-${PGBINDIR}shp2pgsql -W LATIN1 -s 4326 -g the_geom -I ${SHAPEFILEPATH} ${SCHEMA}.${TABLE} | ${PSQLBINDIR}psql -d maps -U ${PGUSER} -q -p ${PGPORT} -f -
+${PGBINDIR}shp2pgsql -W LATIN1 -s ${SRID} -g the_geom -I ${SHP_PATH} ${SCHEMA}.${TABLE} | ${PSQLBINDIR}psql -d maps -U ${PGUSER} -q -p ${PGPORT} -f -
 ${PSQLBINDIR}psql -d maps -U ${PGUSER} -q -p ${PGPORT} -c "
-    INSERT INTO ${SCHEMA}.map_version (table_name, filename) values ('${TABLE}','${SHAPEFILENAME}');
+    INSERT INTO ${SCHEMA}.map_version (table_name, filename) values ('${TABLE}','${SHP_NAME}');
     SELECT AddGeometryColumn('${SCHEMA}','${TABLE}','the_geom_0','4326',(SELECT type FROM public.geometry_columns WHERE f_table_schema='${SCHEMA}' and f_table_name='${TABLE}' and f_geometry_column='the_geom'),2);
     UPDATE ${SCHEMA}.${TABLE} SET the_geom_0=ST_Segmentize(the_geom,0.1);
     CREATE INDEX ${TABLE}_the_geom_0_gist ON ${SCHEMA}.${TABLE} USING gist(the_geom_0);
